Truncate all tables by default

+++ server/datastore/mysql/testing_utils.go
@@ -257,9 +257,10 @@ func CreateNamedMySQLDS(t *testing.T, name string) *Datastore {
 	return ds
 }
 
-// TruncateTables truncates the specified table, in order, using ds.writer.
+// TruncateTables truncates the specified tables, in order, using ds.writer.
 // Note that the order is typically not important because FK checks are
-// disabled while truncating.
+// disabled while truncating. If no table is provided, all tables (except
+// those that are seeded by the SQL schema file) are truncated.
 func TruncateTables(t *testing.T, ds *Datastore, tables ...string) {
 	// those tables are seeded with the schema.sql and as such must not
 	// be truncated - a more precise approach must be used for those, e.g.
@@ -272,12 +273,34 @@ func TruncateTables(t *testing.T, ds *Datastore, tables ...string) {
 	}
 
 	ctx := context.Background()
+
 	require.NoError(t, ds.withTx(ctx, func(tx sqlx.ExtContext) error {
+		var skipSeeded bool
+
+		if len(tables) == 0 {
+			skipSeeded = true
+			sql := `
+      SELECT
+        table_name
+      FROM
+        information_schema.tables
+      WHERE
+        table_schema = database() AND
+        table_type = 'BASE TABLE'
+    `
+			if err := sqlx.SelectContext(ctx, tx, &tables, sql); err != nil {
+				return err
+			}
+		}
+
 		if _, err := tx.ExecContext(ctx, `SET FOREIGN_KEY_CHECKS=0`); err != nil {
 			return err
 		}
 		for _, tbl := range tables {
 			if nonEmptyTables[tbl] {
+				if skipSeeded {
+					continue
+				}
 				return fmt.Errorf("cannot truncate table %s, it contains seed data from schema.sql", tbl)
 			}
 			if _, err := tx.ExecContext(ctx, "TRUNCATE TABLE "+tbl); err != nil {
